requirements.txt
aws-cdk-lib==2.30.0
constructs>=10.0.0,<11.0.0
pydantic==1.9.1
boto3==1.24.22
pytest==6.2.5
